fix(xunitv3-aot): address PR #140 Codex review round 5
Two more real bugs found and fixed:

- Generated TConfig construction could silently resolve to a different,
  more-specific accessible constructor than the one Compono.Generators
  selected and validated - the generated registration lives in the
  consumer's own assembly, so an internal sibling constructor with a more
  specific parameter type (e.g. public TConfig(object) + internal
  TConfig(string)) is genuinely accessible there, and ordinary C# overload
  resolution would pick it over the bare rendered literal argument.
  ConfigProfileBinder never has this problem - ConstructorInfo.Invoke
  invokes the exact constructor it already resolved, no overload
  resolution involved. Fixed by wrapping every rendered constructor
  argument in an explicit cast to the selected constructor's own declared
  parameter type. Proven with a real, JIT-executed test (not just a
  generator snapshot) - each constructor sets a different observable
  value, and the test asserts the public one actually ran.
- EmbeddedTypes' array-element recursion (round 4's own CMP0044 fix)
  crashed with NullReferenceException on a legitimately null argument
  (a bare [Compose<P, C>(null)] targeting a nullable TConfig constructor
  parameter) - confirmed by a direct probe, worse than the "empty array"
  framing suggested. Fixed with an IsNull guard at the top of
  EmbeddedTypes, mirroring the same guard TypedConstantMatcher.Validate/
  TypedConstantLiteralRenderer.Render already had.

New tests: TwoTypeParameterAttribute_ArgumentCastToSelectedConstructorParameterType,
TConfigWithMoreSpecificInternalOverload_UsesSelectedPublicConstructor (real
JIT execution in Compono.XunitV3.Aot.Tests),
TwoTypeParameterAttribute_NullArgumentForNullableParameter_DoesNotCrash.

See docs/plans/0067-compono-xunitv3-aot-profile-support-impl-plan.md's
Notes section for the full account.

+++ test/Compono.Generators.Tests/Snapshots/AotTheoryDataRowRegistrationVerifyTests.TwoTypeParameterAttribute_SpecialFloatingPointValues_RenderAsValidCSharp#TestNamespace.FloatingPointTests.Test_supplies_special_floating_point_values_4fb80701.AotTheoryDataRowRegistration.g.verified.cs
@@ -17,7 +17,7 @@ file static class AotTheoryDataRowRegistration
             disableDiscoveryEnumeration: true,
             factory: static disposalTracker =>
             {
-                var profileConfig = new global::TestNamespace.FloatingPointConfig(double.NaN, double.PositiveInfinity, double.NegativeInfinity, -(double)0, float.NaN);
+                var profileConfig = new global::TestNamespace.FloatingPointConfig((double)double.NaN, (double)double.PositiveInfinity, (double)double.NegativeInfinity, (double)-(double)0, (float)float.NaN);
                 var profileInstance = new global::TestNamespace.FloatingPointProfile(profileConfig);
                 var composer = global::Compono.Composer.Create(b => b.AddProfile(profileInstance));
                 var row = composer.CreateRow(typeof(global::TestNamespace.FloatingPointTests));
